+++ experimental-examples/kitchen-sink/tina/model/location-enum.ts
@@ -1,0 +1,6 @@
+// Define an enum for location
+export enum LocationEnum {
+  Melbourne = 'melbourne',
+  Sydney = 'sydney',
+  Default = 'Default',
+}