+++ packages/@tinacms/cli/src/next/commands/build-command/index.ts
@@ -1,9 +1,10 @@
-import fetch, { Headers } from 'node-fetch'
 import { Command, Option } from 'clipanion'
 import Progress from 'progress'
 import fs from 'fs-extra'
+import crypto from 'crypto'
+import path from 'path'
 import type { ViteDevServer } from 'vite'
-import { buildSchema, Database, FilesystemBridge } from '@tinacms/graphql'
+import { buildSchema, type Database, FilesystemBridge } from '@tinacms/graphql'
 import { ConfigManager } from '../../config-manager'
 import { logger, summary } from '../../../logger'
 import { buildProductionSpa } from './server'
@@ -15,18 +16,19 @@ import {
   getIntrospectionQuery,
 } from 'graphql'
 import { diff } from '@graphql-inspector/core'
-import { IndexStatusResponse, waitForDB } from './waitForDB'
+import { type IndexStatusResponse, waitForDB } from './waitForDB'
 import { createAndInitializeDatabase, createDBServer } from '../../database'
 import { sleepAndCallFunc } from '../../../utils/sleep'
 import { dangerText, linkText, warnText } from '../../../utils/theme'
 import {
-  SearchClient,
+  type SearchClient,
   SearchIndexer,
   TinaCMSSearchIndexClient,
 } from '@tinacms/search'
 import { spin } from '../../../utils/spinner'
 import { createDevServer } from '../dev-command/server'
 import { BaseCommand } from '../baseCommands'
+import { logText } from '../../../utils/theme'
 
 export class BuildCommand extends BaseCommand {
   static paths = [['build']]
@@ -212,6 +214,13 @@ export class BuildCommand extends BaseCommand {
         configManager,
         database,
         codegen.productionUrl
+      )
+      await this.checkTinaSchema(
+        configManager,
+        database,
+        codegen.productionUrl,
+        this.previewName,
+        this.verbose
       )
     }
 
@@ -599,6 +608,75 @@ export class BuildCommand extends BaseCommand {
       }
     }
   }
+
+  async checkTinaSchema(
+    configManager: ConfigManager,
+    database: Database,
+    apiURL: string,
+    previewName: string,
+    verbose: boolean
+  ) {
+    const bar = new Progress(
+      'Checking local Tina Schema matches server. :prog',
+      1
+    )
+    const { config } = configManager
+    const token = config.token
+    const { clientId, branch, isLocalClient, host } = parseURL(apiURL)
+    // Can't check status if we're not using Tina Cloud
+    if (isLocalClient || !host || !clientId || !branch) {
+      if (verbose) {
+        logger.info(logText('Not using Tina Cloud, skipping Tina Schema check'))
+      }
+      return
+    }
+
+    // Get the remote schema from the graphql endpoint
+    const { tinaSchema: remoteTinaSchemaSha } = await fetchSchemaSha({
+      url: `https://${host}/db/${clientId}/${previewName || branch}/schemaSha`,
+      token,
+    })
+
+    if (!remoteTinaSchemaSha) {
+      bar.tick({
+        prog: '❌',
+      })
+      let errorMessage = `The remote Tina schema does not exist. Check indexing for this branch.`
+      if (config?.branch) {
+        errorMessage += `\n\nAdditional info: Branch: ${config.branch}, Client ID: ${config.clientId} `
+      }
+      throw new Error(errorMessage)
+    }
+
+    if (!database.bridge) {
+      throw new Error(`No bridge configured`)
+    }
+    const localTinaSchema = JSON.parse(
+      await database.bridge.get(
+        path.join(database.tinaDirectory, '__generated__', '_schema.json')
+      )
+    )
+    localTinaSchema.version = undefined
+    const localTinaSchemaSha = crypto
+      .createHash('sha256')
+      .update(JSON.stringify(localTinaSchema))
+      .digest('hex')
+
+    if (localTinaSchemaSha === remoteTinaSchemaSha) {
+      bar.tick({
+        prog: '✅',
+      })
+    } else {
+      bar.tick({
+        prog: '❌',
+      })
+      let errorMessage = `The local Tina schema doesn't match the remote Tina schema. Please push up your changes to GitHub to update your remote tina schema.`
+      if (config?.branch) {
+        errorMessage += `\n\nAdditional info: Branch: ${config.branch}, Client ID: ${config.clientId} `
+      }
+      throw new Error(errorMessage)
+    }
+  }
 }
 
 //  This was taken from packages/tinacms/src/unifiedClient/index.ts
@@ -666,7 +744,10 @@ export const fetchRemoteGraphqlSchema = async ({
   if (token) {
     headers.append('X-API-KEY', token)
   }
-  const body = JSON.stringify({ query: getIntrospectionQuery(), variables: {} })
+  const body = JSON.stringify({
+    query: getIntrospectionQuery(),
+    variables: {},
+  })
 
   headers.append('Content-Type', 'application/json')
 
@@ -677,4 +758,24 @@ export const fetchRemoteGraphqlSchema = async ({
   })
   const data = await res.json()
   return data?.data
+}
+
+export const fetchSchemaSha = async ({
+  url,
+  token,
+}: {
+  url: string
+  token?: string
+}) => {
+  const headers = new Headers()
+  if (token) {
+    headers.append('X-API-KEY', token)
+  }
+
+  const res = await fetch(url, {
+    method: 'GET',
+    headers,
+    cache: 'no-cache',
+  })
+  return res.json()
 }