+++ packages/@tinacms/mdx/src/parse/remarkToPlate.ts
@@ -45,32 +45,29 @@ export const remarkToSlate = (
     switch (content.type) {
       case 'table': {
         return {
-          type: 'mdxJsxFlowElement',
-          children: [{ type: 'text', text: '' }],
-          name: 'table',
+          type: 'table',
+          children: content.children.map((tableRow) => {
+            return {
+              type: 'tr',
+              children: tableRow.children.map((tableCell) => {
+                return {
+                  type: 'td',
+                  children: [
+                    {
+                      type: 'p',
+                      children: flatten(
+                        tableCell.children.map((child) =>
+                          phrasingContent(child)
+                        )
+                      ),
+                    },
+                  ],
+                }
+              }),
+            }
+          }),
           props: {
             align: content.align?.filter((item) => !!item),
-            tableRows: content.children.map((child) => {
-              return {
-                tableCells: child.children.map((child) => {
-                  return {
-                    value: {
-                      type: 'root',
-                      children: [
-                        {
-                          type: 'p',
-                          children: flatten(
-                            child.children.map((child) =>
-                              phrasingContent(child)
-                            )
-                          ),
-                        },
-                      ],
-                    },
-                  }
-                }),
-              }
-            }),
           },
         }
       }
@@ -89,7 +86,7 @@ export const remarkToSlate = (
       case 'heading':
         return heading(content)
       case 'code':
-        return code(content)
+        return parseCode(content)
       case 'paragraph':
         return paragraph(content)
       case 'mdxJsxFlowElement':
@@ -285,6 +282,23 @@ export const remarkToSlate = (
           // @ts-ignore
           content.position
         )
+    }
+  }
+
+  const parseCode = (
+    content: Md.Code
+  ): Plate.CodeBlockElement | Plate.MermaidElement => {
+    if (content.lang === 'mermaid') {
+      return mermaid(content)
+    }
+    return code(content)
+  }
+
+  const mermaid = (content: Md.Code): Plate.MermaidElement => {
+    return {
+      type: 'mermaid',
+      value: content.value,
+      children: [{ type: 'text', text: '' }],
     }
   }