+++ packages/@tinacms/schema-tools/src/schema/TinaSchema.ts
@@ -68,6 +68,22 @@ export class TinaSchema {
       collectionNames.includes(collection.name)
     )
   }
+
+  public findReferences(name: string) {
+    const result: Record<string, { path: string[]; field: TinaField }[]> = {}
+    this.walkFields(({ field, collection: c, path }) => {
+      if (field.type === 'reference') {
+        if (field.collections.includes(name)) {
+          if (result[c.name] === undefined) {
+            result[c.name] = []
+          }
+          result[c.name].push({ path, field })
+        }
+      }
+    })
+    return result
+  }
+
   public getCollection = (collectionName: string): Collection<true> => {
     const collection = this.schema.collections.find(
       (collection) => collection.name === collectionName
@@ -282,7 +298,6 @@ export class TinaSchema {
     collection: Collectable
   ) => {
     const accumulator: { [key: string]: unknown } = {}
-    // biome-ignore lint/complexity/noForEach: <explanation>
     Object.entries(payload).forEach(([key, value]) => {
       if (typeof collection.fields === 'string') {
         throw new Error('Global templates not supported')