+++ packages/@tinacms/schema-tools/src/types/index.ts
@@ -7,6 +7,10 @@ type Meta = {
   error?: any
 }
 
+// This type is used in collectionFilter for reference field, it represent the datatype of the field in the referenced collection that will be used to do the filter
+// This type is extendable if we need to support more type for reference selection filter (e.g. boolean)
+type FilterValue = string[] | string
+
 type Component<Type, List> = (props: {
   field: TinaField & { namespace: string[] }
   input: {
@@ -215,7 +219,13 @@ export type ImageField = (
 
 type ReferenceFieldOptions = {
   optionComponent?: OptionComponent
+  /**
+   * @deprecated use `collectionFilter` instead as experimental___Filter will be removed in a future release
+   */
   experimental___filter?: (list: Array<any>, searchQuery: string) => Array<any>
+  collectionFilter?:
+    | Record<string, Record<string, FilterValue>>
+    | (() => Record<string, Record<string, FilterValue>>)
 }
 
 type OptionComponent<P = Record<string, unknown>, S = Document['_sys']> = (
@@ -250,7 +260,7 @@ export type PasswordField = (
     type: 'password'
   }
 
-type toolbarItemName =
+type ToolbarOverrideType =
   | 'heading'
   | 'link'
   | 'image'
@@ -263,6 +273,8 @@ type toolbarItemName =
   | 'italic'
   | 'raw'
   | 'embed'
+  | 'mermaid'
+  | 'table'
 type RichTextAst = { type: 'root'; children: Record<string, unknown>[] }
 export type RichTextField<WithNamespace extends boolean = false> = (
   | FieldGeneric<RichTextAst, undefined>
@@ -277,7 +289,7 @@ export type RichTextField<WithNamespace extends boolean = false> = (
      * will be stored as frontmatter
      */
     isBody?: boolean
-    toolbarOverride?: toolbarItemName[]
+    toolbarOverride?: ToolbarOverrideType[]
     templates?: RichTextTemplate<WithNamespace>[]
     /**
      * By default, Tina parses markdown with MDX, this is a more strict parser
@@ -818,6 +830,7 @@ type Document = {
     relativePath: string
     filename: string
     extension: string
+    hasReferences?: boolean
   }
 }
 export interface UICollection<Form = any, CMS = any, TinaForm = any> {
@@ -842,6 +855,15 @@ export interface UICollection<Form = any, CMS = any, TinaForm = any> {
      * When set to `true`, editors won't be able to modify the filename
      */
     readonly?: boolean
+    /**
+     * When set to `true`, the filename will be shown first in the form
+     * @default false
+     */
+    showFirst?: boolean
+    /**
+     * Sets the description for the filename field
+     */
+    description?: string
   }
   /**
    * Determines whether or not this collection can accept new docments